# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API para el sistema interno de Mariana Textil.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Inicia sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Sesión creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Obtiene la sesión actual
      responses:
        "200":
          description: Usuario autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Sesión no válida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Demasiados intentos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Role:
      type: string
      enum: [ADMIN, CAJA, INVENTARIOS, BODEGA]
    LocationType:
      type: string
      enum: [TIENDA, BODEGA, TRANSITO, EXTERNO]
    Location:
      type: object
      required: [id, nombre, tipo, activa, esSistema]
      properties:
        id: { type: number }
        nombre: { type: string }
        tipo:
          $ref: "#/components/schemas/LocationType"
        activa: { type: boolean }
        esSistema: { type: boolean }
...
    AlcanceConsulta:
      type: string
      enum: [PROPIA, TODAS]
    ModulePermission:
      type: object
      required: [modulo, puedeVer, puedeCrear, puedeEditar, puedeAutorizar]
      properties:
        modulo: { type: string }
        puedeVer: { type: boolean }
        puedeCrear: { type: boolean }
        puedeEditar: { type: boolean }
        puedeAutorizar: { type: boolean }
...
    PermissionMatrix:
      type: array
      description: Lista de permisos efectivos del usuario (uno por módulo)
      items:
        $ref: "#/components/schemas/ModulePermission"
...
    CurrentUser:
      type: object
      required: [id, nombre, usuario, rol, ubicacion, alcanceConsulta, permisos]
      properties:
        id: { type: number }
        nombre: { type: string }
        usuario: { type: string }
        rol:
          $ref: "#/components/schemas/Role"
        ubicacion:
          oneOf:
            - $ref: "#/components/schemas/Location"
            - type: "null"
        alcanceConsulta:
          $ref: "#/components/schemas/AlcanceConsulta"
        permisos:
          $ref: "#/components/schemas/PermissionMatrix"
...
    LoginInput:
      type: object
      required: [usuario, password]
      properties:
        usuario: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 128 }
...